+++ HISTORY.rst
@@ -7,5 +7,6 @@ History
 * Added commands for applying fixities to list of nodes (e.g. `o3.Fix2DOFMulti`), and for equal DOF command
 * Added function for generating a grid of nodes `build_regular_node_mesh`
 * Added option for compressing the output of a opy file by applying for loops
+* Added `add_fixity_to_dof` to try to apply fixity but not fail if fixity already existing